refactor 상수 변수화, 테스트 클래스 분리, 빈값체크 메소드 수정
